Facilities to add code to the editor via builder functions
I've passed in the `model` into the AST Context which seems to be taking
more responsibility now (which I guess is ok).

The split between AST and editor is a bit annoying right now..

AstContext has a `textFactory` function now, which takes a builder and
returns text of the given node.